+++ lib/features/dashboard/presentation/widgets/dashboard_map.dart
@@ -96,7 +96,6 @@ class DashBoardMap extends StatelessWidget {
             myLocationEnabled: true,
             myLocationButtonEnabled: true,
             markers: markers,
-            onMapCreated: (GoogleMapController controller) {},
           ),
         ),
       ),
